% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{aqDropHours}
\alias{aqDropHours}
\title{Drops data of several hours, delegates on to aqDropHour}
\usage{
aqDropHours(x, hours)
}
\arguments{
\item{x}{the input xts data set}

\item{hours}{the vector of hours to drop}
}
\value{
the resulting data set
}
\description{
Drops data of several hours, delegates on to aqDropHour
}
