\name{aqEnableDebugMessages}
\alias{aqEnableDebugMessages}
\title{Enable debug messages}
\description{	
  Enables debug messages in messaging related C-parts of AQ-R. Debug messages provide a variety of additional information, such as the raw messages received. 
}
\usage{
	aqEnableDebugMessages()
}
\keyword{messaging}
