\name{aqHourlyStat}
\alias{aqHourlyStat}
\title{Applies a function to a group of data per hour}
\description{
	This function groups data by the hour and applies a function to it. 
}
\usage{
	aqHourlyStat(x, f = mean)
}
\arguments{
\item{x}{an xts data set}
\item{f}{a function which should be applied hour-wise}
}
\keyword{utilities}
