\name{aqInitMessaging}
\alias{aqInitMessaging}
\title{Initializes the messaging layer}
\description{	
  The messaging layer has to be initialized. 
}
\usage{
	aqInitMessaging(host = "localhost", port = 61618)
}

\arguments{
\item{host}{the STOMP host}
\item{port}{the STOMP port}
}
\keyword{messaging}
\keyword{initialization}
