\name{aqLoadXtsFromCsv}
\alias{aqLoadXtsFromCsv}
\title{Loads a XTS object from CSV, to be used with our aqSaveXtsToCsv function. This method
assumes that the fie's first column contains an interpretable timestmap.}
\usage{
aqLoadXtsFromCsv(filename)
}
\arguments{
  \item{filename}{the csv file which to load as XTS.}
}
\value{
an XTS object
}
\description{
Implementation in progress (16 Feb 2014)
}

