% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist_data.R
\name{aqSaveXtsToCsv}
\alias{aqSaveXtsToCsv}
\title{Saves an XTS object to csv file.}
\usage{
aqSaveXtsToCsv(filename, historyXts)
}
\arguments{
\item{filename}{where to save data to}

\item{historyXts}{the input xts object}
}
\description{
Saves an XTS object to csv file.
}
