% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist_data.R
\name{aqStoreSeriesField}
\alias{aqStoreSeriesField}
\title{Stores one series field to an AQ Master Server, typicall called from aqStoreSeries. 
This function assumes that data is either a zoo object, or that is a matrix with two columns where the first column contains a time series index in NANOSECONDS(!!!)}
\usage{
aqStoreSeriesField(
  seriesId,
  fieldId,
  freq,
  data,
  con = aqInit(),
  silent = FALSE,
  writeKey = "unset"
)
}
\arguments{
\item{seriesId}{a series name}

\item{fieldId}{the field ID of this data series}

\item{freq}{a frequency string, must be one of AQ's suported enum names}

\item{data}{the data as XTS object}

\item{con}{a connection object, will be initialized by aqInit by default}

\item{silent}{whether  it should print storage diagnostics.}

\item{writeKey}{the write key, unset by default}
}
\description{
Stores one series field to an AQ Master Server, typicall called from aqStoreSeries. 
This function assumes that data is either a zoo object, or that is a matrix with two columns where the first column contains a time series index in NANOSECONDS(!!!)
}
