\name{aqUnsubscribeChannel}
\alias{aqUnsubscribeChannel}
\title{Unsubscribe from a messaging channel}
\description{	
  Unsubscribes from a messaging channel. The bridge will send an unsubscribe command to the messaging server. 
}
\usage{
	aqUnsubscribeChannel(channel)
}
\arguments{
\item{channel}{one channel name}
}
\keyword{messaging}
