\name{aqWaitForData}
\alias{aqWaitForData}
\title{Wait for data at the bridge}
\description{	
  A blocking call that waits for data at the messaging bridge. The call will return a list of channels for which data is available, once the bridge contains data. 
}
\usage{
	aqWaitForData()
}
\keyword{messaging}
