% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{backadjust}
\alias{backadjust}
\title{Backadjusts data, expects yahoo xts structure: 
column 1-4: ohlc, 5: vol, 6: backadjusted close price}
\usage{
backadjust(yahooOhlc)
}
\description{
Backadjusts data, expects yahoo xts structure: 
column 1-4: ohlc, 5: vol, 6: backadjusted close price
}
