% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{d8}
\alias{d8}
\title{Formats an input POSIX date to Date8 format (YYYYMMDD). This function is a convenience function.}
\usage{
d8(dateObj)
}
\arguments{
\item{dateObj}{a POSIX date object}
}
\value{
an integer YYYYMMDD representation
}
\description{
Formats an input POSIX date to Date8 format (YYYYMMDD). This function is a convenience function.
}

