% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{daysAgo}
\alias{daysAgo}
\title{returns the date X days ago (30 days by default)}
\usage{
daysAgo(d = 30)
}
\arguments{
\item{d}{amount of days to shift, default: 30}
}
\value{
a POSIXlt object
}
\description{
returns the date X days ago (30 days by default)
}

