% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{normalize}
\alias{normalize}
\title{Normalizes an input matrix along certain dimensions to a range from 0 to 1. Retains the matrix class, such as XTS, ZOO, etc.}
\usage{
normalize(inputMatrix, normRange = 1)
}
\arguments{
\item{inputMatrix}{an input matrix with numbers, NAs not allowed.}

\item{normRange}{specifies by which range to normalize, 0 = entire matrix, 1 = per row, 2 = per column}
}
\value{
a normalized copy of the input matrix
}
\description{
Normalizes an input matrix along certain dimensions to a range from 0 to 1. Retains the matrix class, such as XTS, ZOO, etc.
}

