% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtest.R
\name{approximateSLTP}
\alias{approximateSLTP}
\title{Applies an approximated StopLoss/TakeProfit strategy on an incoming pnl series.}
\usage{
approximateSLTP(
  high,
  low,
  close,
  takeProfit,
  stopLoss,
  runningPosition,
  messages = FALSE
)
}
\value{
the rewritten pnl series.
}
\description{
Applies an approximated StopLoss/TakeProfit strategy on an incoming pnl series.
}
