\name{aqHourIndex}
\alias{aqHourIndex}
\title{Returns for an XTS input list the hour index per element.}
\usage{
aqHourIndex(xtsSeries)
}
\arguments{
  \item{xtsSeries}{the input object of type XTS.}
}
\value{
a vector of the same length as xtsSeries, containing the
hour
}
\description{
Returns for an XTS input list the hour index per element.
}

