% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aq_config.R
\name{aqInit}
\alias{aqInit}
\title{This method builds an object that carries necessary configuration values. 
The resulting object is a list, which you can modify outside of this function.
Currently contains tsHost, tsPort, openField,closeField, highField, lowField
and volField.}
\usage{
aqInit()
}
\value{
This function returns a plain list with configuration settings.
}
\description{
This method builds an object that carries necessary configuration values. 
The resulting object is a list, which you can modify outside of this function.
Currently contains tsHost, tsPort, openField,closeField, highField, lowField
and volField.
}
