% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist_data.R
\name{aqLoadOHLC}
\alias{aqLoadOHLC}
\title{Loads OHLC from an AQ Master Server}
\usage{
aqLoadOHLC(
  seriesId,
  freq,
  startDate,
  endDate,
  con = aqInit(),
  useCache = FALSE,
  cacheDir = getwd()
)
}
\arguments{
\item{seriesId}{a series ID}

\item{freq}{frequency in enumeration form, f.e. HOURS_1, MINUTES_1}

\item{startDate}{a Date8}

\item{endDate}{a Date8}

\item{con}{a fully initialized connection definition}

\item{useCache}{a boolean that says whether you want use and cache data}

\item{cacheDir}{a directory name that will be used for caching if enabled}
}
\value{
a XTS object
}
\description{
Loads OHLC from an AQ Master Server
}
