% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist_data.R
\name{aqLoadXtsFromCsv}
\alias{aqLoadXtsFromCsv}
\title{Loads a XTS object from CSV, to be used with our aqSaveXtsToCsv function. This method 
assumes that the fie's first column contains an interpretable timestmap. 
  
Implementation in progress (16 Feb 2014)}
\usage{
aqLoadXtsFromCsv(filename)
}
\arguments{
\item{filename}{the csv file which to load as XTS.}
}
\value{
an XTS object
}
\description{
Loads a XTS object from CSV, to be used with our aqSaveXtsToCsv function. This method 
assumes that the fie's first column contains an interpretable timestmap. 
  
Implementation in progress (16 Feb 2014)
}
